pyproject.toml
[project]
name = "unajyu"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "requests>=2.31.0",
    "discord.py>=2.3.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
